+++ Mark1Engine/Vector2.cs
@@ -9,15 +9,15 @@ namespace chess.Mark1Engine
     public class Vector2 : IEquatable<Vector2>
     {
 
-        public float x;
-        public float y;
+        public int x;
+        public int y;
 
         public Vector2()
         {
 
         }
 
-        public Vector2(float x, float y)
+        public Vector2(int x, int y)
         {
             this.x = x;
             this.y = y;
